rename decorator.py -> cli.py
