+++ localstack/utils/analytics/cli.py
